New ASCII art for Linux mint (like Manjaro)
I've installed neofetch for the first time and I was really amazed to see the ASCII art of the Manjaro distro. But I was on Linux mint and I wanted to make my terminal look as good as the manjaro one looks so I designed the ASCII art for Linux mint as like the manjaro one. 
unnecessary note: I'm not a programmer but somehow I figured out how the code works and I copied an existing code block from the neofetch file and and used it as a template .